+++ script/core/reference.lua
@@ -2,6 +2,7 @@ local guide      = require 'parser.guide'
 local files      = require 'files'
 local vm         = require 'vm'
 local findSource = require 'core.find-source'
+local jumpSource = require 'core.jump-source'
 
 local function sortResults(results)
     -- 先按照顺序排序
@@ -132,6 +133,7 @@ return function (uri, position)
     end
 
     sortResults(results)
+    jumpSource(results)
 
     return results
 end